TDD of DB initialization - 1 of 2 tests working
Following TDD, test for initialization of DB (registering JDBC driver,
connecting to DB, creating FILE and PEER tables) is implemented. Only
tests to see if after creation, a valid connection exists. Tables and
constraints will be tested in separate methods. Also, test for multiple
connections created. Not working at the moment and needs to be changed.
Code for these tests has been implemented in DatabaseManager, which now
uses JDBCDriverException and MultipleConnectionException for more
specific error handling that may be done in the future.