Added basic classes to rep. Peers and PeerFiles
Added basic versions of classes that corresond to the Peers in the
network and the files that they are sharing via PeerFiles. Note that the
Peer class has little relation to the (future) PeerComputer class; the
latter is used by the application to represent the user's machine as a
peer (allows other peers to send messages to them, downloaded files,
etc) while the former simply is a class corresponding to each other peer
in the network, containing the basic info (corresponding to that in the
DB). Similarly, PeerFiles have the exact same data as the Files in the
DB. Used for convenience, and will eventually be altered to work with
JPA (Java Persistence API).