Added shells of class/interface DTManager & Sender
Created DataTransferManager class (and test class), as well as Sender
interface. DataTransferManager will be in control of all data transfer
(sending to other machines & receiving for other machines). The absolute
minimum has been made for this class at this point (class & test wise)
because this manager will rely on a Sender and Receiver instances, which
are not yet implemented. Shell of Sender interface has been created.
Note that DataTransferManager will not be tied to a specific Sender or
Receiver to allow for easy integration of different UDP protocols (e.g.
(& i.e. for now) stop-&-wait protocol (byte-by-byte), selective repeat
(pipelined protocol))