Major progress on Sender & StopAndWaitSender
StopAndWaitSender has a good deal of functionality implemented and
tested. It can now get appropriately sized message chunks, apply
transport layer headers and potentially send packets to a given peer.
The parent abstract class Sender has also been changed and now follows
the Template design pattern; the sendMessage method as it currently is
establishes a basic means by which all Sender subclasses will send by.
However, this method is not tested and does not account for ACK
checking. More functionality is expected to be delegated to the Sender
class, such as sending packets, but this will be done when needed.
Beyond this, an exception class has been created for sending packets and
a NetworkCommunicator class has been created to hold data and
functionality shared by Senders and Receivers. It's use is minimal now,
holding constants used by Senders.